backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
huggingface_hub
python-multipart
